+++ lib/road.py
@@ -6,9 +6,10 @@ class Road:
     def __init__(self, display, car):
         self.display = display
         self.car = car
+        self.speed = 5
         self.rect = pygame.Rect((500 * display.scr_w, 0, 920 * display.scr_w, display.screen_height))
-        self.obstacle1 = Obstacles(display, random.randrange(500, 1420 - 200), car)
-        self.obstacle2 = Obstacles(display, random.randrange(500, 1420 - 200), car)
+        self.obstacle1 = Obstacles(display, random.randrange(500, 1420 - 200), car, self.speed)
+        self.obstacle2 = Obstacles(display, random.randrange(500, 1420 - 200), car, self.speed)
 
     def generate(self):
         if self.obstacle1.rect.top < self.display.screen_height or self.obstacle2.rect.top < self.display.screen_height:
@@ -18,22 +19,22 @@ class Road:
             self.obstacle1.draw(self.display.screen)
             self.obstacle2.draw(self.display.screen)
         else:
-            self.obstacle1 = Obstacles(self.display, random.randrange(500, 1420 - 200), self.car)
-            self.obstacle2 = Obstacles(self.display, random.randrange(500, 1420 - 200), self.car)
+            self.obstacle1 = Obstacles(self.display, random.randrange(500, 1420 - 200), self.car, self.speed)
+            self.obstacle2 = Obstacles(self.display, random.randrange(500, 1420 - 200), self.car, self.speed)
 
 
 class Obstacles:
-    def __init__(self, display, x, target):
+    def __init__(self, display, x, target, speed):
         self.display = display
         self.rect = pygame.Rect((x, 0 - 200 * display.scr_h, 200 * display.scr_w, 200 * display.scr_h))
         self.target = target
+        self.speed = speed
 
     def move(self):
-        SPEED = 7
         # update obstacle position
-        self.rect.y += SPEED
+        self.rect.y += self.speed
         if self.target.rect.colliderect(self.rect) and self.target.hit_cooldown <= 0:
-            self.target.health -= 20
+            self.target.health -= 50
             self.target.hit_cooldown = 80
 
     def draw(self, surface):
